+++ src/main/java/com/ixayda/iam/user/LoginIdentifierType.java
@@ -1,0 +1,11 @@
+package com.ixayda.iam.user;
+
+public enum LoginIdentifierType {
+
+	USERNAME,
+
+	EMAIL,
+
+	PHONE
+
+}